Refactor the XState compiler
This refactor breaks the XState compiler into smaller chunks, making it
easier to make additions/changes.
