+++ src/core/identifier.h
@@ -1,7 +1,4 @@
-#ifndef LUCY_IDENTIFIER_H_
-#define LUCY_IDENTIFIER_H_
+#pragma once
 
 void identifier_init();
 int is_valid_identifier_char(char);
-
-#endif
