+++ src/core/js_builder.c
@@ -1,4 +1,3 @@
-#include <stdio.h> // todo remove
 #include <stdlib.h>
 #include <string.h>
 #include <stdbool.h>
