+++ src/core/xstate/assignment.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../js_builder.h"
+#include "../node.h"
+#include "core.h"
+
+void xs_enter_assignment(PrintState*, JSBuilder*, Node*);
