+++ src/core/xstate/import.c
@@ -1,0 +1,37 @@
+#include <stdio.h>
+#include "../js_builder.h"
+#include "../node.h"
+#include "core.h"
+
+void xs_enter_import(PrintState* state, JSBuilder* jsb, Node* node) {
+  js_builder_add_str(jsb, "import ");
+
+  ImportNode *import_node = (ImportNode*)node;
+  Node* child = node->child;
+  bool multiple = false;
+
+  if(child == NULL) {
+    printf("TODO add support for imports with no specifiers\n");
+    return;
+  }
+
+  js_builder_add_str(jsb, "{ ");
+
+  while(child != NULL) {
+    ImportSpecifier *specifier = (ImportSpecifier*)child;
+
+    if(multiple) {
+      js_builder_add_str(jsb, ", ");
+    }
+
+    js_builder_add_str(jsb, specifier->imported);
+    // TODO support local
+
+    child = child->next;
+    multiple = true;
+  }
+
+  js_builder_add_str(jsb, " } from ");
+  js_builder_add_str(jsb, import_node->from);
+  js_builder_add_str(jsb, ";\n");
+}
