+++ src/core/xstate/import.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../js_builder.h"
+#include "../node.h"
+#include "core.h"
+
+void xs_enter_import(PrintState*, JSBuilder*, Node*);
