+++ src/core/xstate/invoke.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../js_builder.h"
+#include "../node.h"
+#include "core.h"
+
+void xs_compile_invoke(PrintState*, JSBuilder*, InvokeNode*);
