+++ src/core/xstate/machine.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "../js_builder.h"
+#include "../node.h"
+#include "core.h"
+
+void xs_add_machine_binding_name(JSBuilder*, MachineNode*);
+void xs_enter_machine(PrintState*, JSBuilder*, Node*);
+void xs_exit_machine(PrintState*, JSBuilder*, Node*);
